emp salary prediction/requirements.txt
# requirements.txt
pandas
scikit-learn
streamlit
joblib
google-generativeai 